Replace print() with python logger.
This commit fixes bug on MacOS: the app crashed when pressing 'Recognize'
button if selected files contained non-ascii characters in its path. I
was able to reproduce the bug only when launching the app bundle from
Finder. Being launched from Terminal, the app worked as expected.

The app crashed with the following message and traceback:
Traceback (most recent call last):
  File "godr/frontend/app_ui.py", line 114
    print('Files are {}'.format(self.selected_files))
UnicodeEncodeError: 'ascii' codec can't encode character in position <...>

It turned out that in the described setup sys.stdout.encoding == 'ascii',
so it was not able to print russian letters. Switching to logging module
instead of print solved the issue.